fix: do not hash anything within this crate
`new_from_prefix()` is the only function that does hashing of the data.
For a clear separation of concerns, it is changed to take a hash instead.
The hashing should be done before a CID is created.

BREAKING CHANGE: `new_from_prefix()` takes a hash intead of data

Prior to this change `new_from_prefix()` took some data that needs to
be hashed as input. With this change, you need to hash the data before
this call and then pass the hash digest in.