fix: follow same pattern as multihash
In `multihash` there is a `Multihash` type alias exported with the
size of the codec table. The type generic over the size is exported
as `MultihashGeneric`.

BREAKING CHANGE: Follow the same pattern as multihash and call
the CID that is generic over the digest size `CidGeneric`, instead
of making the whole `cid` module public.

Also don't export multihash `Size` and `U64` directly anymore. You
have access to it through the `multihash` re-export.