fix: `varint_read_u64` panics on unwrap in `no_std` environment (#145)
The goal of this pull request is to refactor `unwrap()` to return the `u64` value or an`Error:VarIntDecodeError` in the `no_std` environment.

This error was discovered in fuzz testing.

Tests were refactored to include a new `no_std_tests` module, in order to test the fix. 
`cargo test --no-default-features` will execute the `no_std` test.